Start decoupling the decoding and relection
Although this doesn't provide immediate benefit, the intent is to make
the data decoder available separately in a future commit for public use.
