+++ internal/decoder/decoder.go
@@ -3,18 +3,14 @@ package decoder
 
 import (
 	"encoding/binary"
-	"errors"
-	"fmt"
 	"math"
 	"math/big"
-	"reflect"
-	"sync"
 
 	"github.com/oschwald/maxminddb-golang/v2/internal/mmdberrors"
 )
 
-// Decoder is a decoder for the MMDB data section.
-type Decoder struct {
+// DataDecoder is a decoder for the MMDB data section.
+type DataDecoder struct {
 	buffer []byte
 }
 
@@ -46,47 +42,12 @@ const (
 	maximumDataStructureDepth = 512
 )
 
-// New creates a [Decoder].
-func New(buffer []byte) Decoder {
-	return Decoder{buffer: buffer}
+// NewDataDecoder creates a [DataDecoder].
+func NewDataDecoder(buffer []byte) DataDecoder {
+	return DataDecoder{buffer: buffer}
 }
 
-// Decode decodes the data value at offset and stores it in the value
-// pointed at by v.
-func (d *Decoder) Decode(offset uint, v any) error {
-	rv := reflect.ValueOf(v)
-	if rv.Kind() != reflect.Ptr || rv.IsNil() {
-		return errors.New("result param must be a pointer")
-	}
-
-	if dser, ok := v.(deserializer); ok {
-		_, err := d.decodeToDeserializer(offset, dser, 0, false)
-		return err
-	}
-
-	_, err := d.decode(offset, rv, 0)
-	return err
-}
-
-func (d *Decoder) decode(offset uint, result reflect.Value, depth int) (uint, error) {
-	if depth > maximumDataStructureDepth {
-		return 0, mmdberrors.NewInvalidDatabaseError(
-			"exceeded maximum data structure depth; database is likely corrupt",
-		)
-	}
-	typeNum, size, newOffset, err := d.decodeCtrlData(offset)
-	if err != nil {
-		return 0, err
-	}
-
-	if typeNum != _Pointer && result.Kind() == reflect.Uintptr {
-		result.Set(reflect.ValueOf(uintptr(offset)))
-		return d.nextValueOffset(offset, 1)
-	}
-	return d.decodeFromType(typeNum, size, newOffset, result, depth+1)
-}
-
-func (d *Decoder) decodeToDeserializer(
+func (d *DataDecoder) decodeToDeserializer(
 	offset uint,
 	dser deserializer,
 	depth int,
@@ -116,98 +77,7 @@ func (d *Decoder) decodeToDeserializer(
 	return d.decodeFromTypeToDeserializer(typeNum, size, newOffset, dser, depth+1)
 }
 
-// DecodePath decodes the data value at offset and stores the value assocated
-// with the path in the value pointed at by v.
-func (d *Decoder) DecodePath(
-	offset uint,
-	path []any,
-	v any,
-) error {
-	result := reflect.ValueOf(v)
-	if result.Kind() != reflect.Ptr || result.IsNil() {
-		return errors.New("result param must be a pointer")
-	}
-
-PATH:
-	for i, v := range path {
-		var (
-			typeNum dataType
-			size    uint
-			err     error
-		)
-		typeNum, size, offset, err = d.decodeCtrlData(offset)
-		if err != nil {
-			return err
-		}
-
-		if typeNum == _Pointer {
-			pointer, _, err := d.decodePointer(size, offset)
-			if err != nil {
-				return err
-			}
-
-			typeNum, size, offset, err = d.decodeCtrlData(pointer)
-			if err != nil {
-				return err
-			}
-		}
-
-		switch v := v.(type) {
-		case string:
-			// We are expecting a map
-			if typeNum != _Map {
-				// XXX - use type names in errors.
-				return fmt.Errorf("expected a map for %s but found %d", v, typeNum)
-			}
-			for range size {
-				var key []byte
-				key, offset, err = d.decodeKey(offset)
-				if err != nil {
-					return err
-				}
-				if string(key) == v {
-					continue PATH
-				}
-				offset, err = d.nextValueOffset(offset, 1)
-				if err != nil {
-					return err
-				}
-			}
-			// Not found. Maybe return a boolean?
-			return nil
-		case int:
-			// We are expecting an array
-			if typeNum != _Slice {
-				// XXX - use type names in errors.
-				return fmt.Errorf("expected a slice for %d but found %d", v, typeNum)
-			}
-			var i uint
-			if v < 0 {
-				if size < uint(-v) {
-					// Slice is smaller than negative index, not found
-					return nil
-				}
-				i = size - uint(-v)
-			} else {
-				if size <= uint(v) {
-					// Slice is smaller than index, not found
-					return nil
-				}
-				i = uint(v)
-			}
-			offset, err = d.nextValueOffset(offset, i)
-			if err != nil {
-				return err
-			}
-		default:
-			return fmt.Errorf("unexpected type for %d value in path, %v: %T", i, v, v)
-		}
-	}
-	_, err := d.decode(offset, result, len(path))
-	return err
-}
-
-func (d *Decoder) decodeCtrlData(offset uint) (dataType, uint, uint, error) {
+func (d *DataDecoder) decodeCtrlData(offset uint) (dataType, uint, uint, error) {
 	newOffset := offset + 1
 	if offset >= uint(len(d.buffer)) {
 		return 0, 0, 0, mmdberrors.NewOffsetError()
@@ -228,7 +98,7 @@ func (d *Decoder) decodeCtrlData(offset uint) (dataType, uint, uint, error) {
 	return typeNum, size, newOffset, err
 }
 
-func (d *Decoder) sizeFromCtrlByte(
+func (d *DataDecoder) sizeFromCtrlByte(
 	ctrlByte byte,
 	offset uint,
 	typeNum dataType,
@@ -263,56 +133,7 @@ func (d *Decoder) sizeFromCtrlByte(
 	return size, newOffset, nil
 }
 
-func (d *Decoder) decodeFromType(
-	dtype dataType,
-	size uint,
-	offset uint,
-	result reflect.Value,
-	depth int,
-) (uint, error) {
-	result = indirect(result)
-
-	// For these types, size has a special meaning
-	switch dtype {
-	case _Bool:
-		return unmarshalBool(size, offset, result)
-	case _Map:
-		return d.unmarshalMap(size, offset, result, depth)
-	case _Pointer:
-		return d.unmarshalPointer(size, offset, result, depth)
-	case _Slice:
-		return d.unmarshalSlice(size, offset, result, depth)
-	}
-
-	// For the remaining types, size is the byte size
-	if offset+size > uint(len(d.buffer)) {
-		return 0, mmdberrors.NewOffsetError()
-	}
-	switch dtype {
-	case _Bytes:
-		return d.unmarshalBytes(size, offset, result)
-	case _Float32:
-		return d.unmarshalFloat32(size, offset, result)
-	case _Float64:
-		return d.unmarshalFloat64(size, offset, result)
-	case _Int32:
-		return d.unmarshalInt32(size, offset, result)
-	case _String:
-		return d.unmarshalString(size, offset, result)
-	case _Uint16:
-		return d.unmarshalUint(size, offset, result, 16)
-	case _Uint32:
-		return d.unmarshalUint(size, offset, result, 32)
-	case _Uint64:
-		return d.unmarshalUint(size, offset, result, 64)
-	case _Uint128:
-		return d.unmarshalUint128(size, offset, result)
-	default:
-		return 0, mmdberrors.NewInvalidDatabaseError("unknown type: %d", dtype)
-	}
-}
-
-func (d *Decoder) decodeFromTypeToDeserializer(
+func (d *DataDecoder) decodeFromTypeToDeserializer(
 	dtype dataType,
 	size uint,
 	offset uint,
@@ -374,326 +195,30 @@ func (d *Decoder) decodeFromTypeToDeserializer(
 	}
 }
 
-func unmarshalBool(size, offset uint, result reflect.Value) (uint, error) {
-	if size > 1 {
-		return 0, mmdberrors.NewInvalidDatabaseError(
-			"the MaxMind DB file's data section contains bad data (bool size of %v)",
-			size,
-		)
-	}
-	value, newOffset := decodeBool(size, offset)
-
-	switch result.Kind() {
-	case reflect.Bool:
-		result.SetBool(value)
-		return newOffset, nil
-	case reflect.Interface:
-		if result.NumMethod() == 0 {
-			result.Set(reflect.ValueOf(value))
-			return newOffset, nil
-		}
-	}
-	return newOffset, mmdberrors.NewUnmarshalTypeError(value, result.Type())
-}
-
-// indirect follows pointers and create values as necessary. This is
-// heavily based on encoding/json as my original version had a subtle
-// bug. This method should be considered to be licensed under
-// https://golang.org/LICENSE
-func indirect(result reflect.Value) reflect.Value {
-	for {
-		// Load value from interface, but only if the result will be
-		// usefully addressable.
-		if result.Kind() == reflect.Interface && !result.IsNil() {
-			e := result.Elem()
-			if e.Kind() == reflect.Ptr && !e.IsNil() {
-				result = e
-				continue
-			}
-		}
-
-		if result.Kind() != reflect.Ptr {
-			break
-		}
-
-		if result.IsNil() {
-			result.Set(reflect.New(result.Type().Elem()))
-		}
-
-		result = result.Elem()
-	}
-	return result
-}
-
-var sliceType = reflect.TypeOf([]byte{})
-
-func (d *Decoder) unmarshalBytes(size, offset uint, result reflect.Value) (uint, error) {
-	value, newOffset := d.decodeBytes(size, offset)
-
-	switch result.Kind() {
-	case reflect.Slice:
-		if result.Type() == sliceType {
-			result.SetBytes(value)
-			return newOffset, nil
-		}
-	case reflect.Interface:
-		if result.NumMethod() == 0 {
-			result.Set(reflect.ValueOf(value))
-			return newOffset, nil
-		}
-	}
-	return newOffset, mmdberrors.NewUnmarshalTypeError(value, result.Type())
-}
-
-func (d *Decoder) unmarshalFloat32(size, offset uint, result reflect.Value) (uint, error) {
-	if size != 4 {
-		return 0, mmdberrors.NewInvalidDatabaseError(
-			"the MaxMind DB file's data section contains bad data (float32 size of %v)",
-			size,
-		)
-	}
-	value, newOffset := d.decodeFloat32(size, offset)
-
-	switch result.Kind() {
-	case reflect.Float32, reflect.Float64:
-		result.SetFloat(float64(value))
-		return newOffset, nil
-	case reflect.Interface:
-		if result.NumMethod() == 0 {
-			result.Set(reflect.ValueOf(value))
-			return newOffset, nil
-		}
-	}
-	return newOffset, mmdberrors.NewUnmarshalTypeError(value, result.Type())
-}
-
-func (d *Decoder) unmarshalFloat64(size, offset uint, result reflect.Value) (uint, error) {
-	if size != 8 {
-		return 0, mmdberrors.NewInvalidDatabaseError(
-			"the MaxMind DB file's data section contains bad data (float 64 size of %v)",
-			size,
-		)
-	}
-	value, newOffset := d.decodeFloat64(size, offset)
-
-	switch result.Kind() {
-	case reflect.Float32, reflect.Float64:
-		if result.OverflowFloat(value) {
-			return 0, mmdberrors.NewUnmarshalTypeError(value, result.Type())
-		}
-		result.SetFloat(value)
-		return newOffset, nil
-	case reflect.Interface:
-		if result.NumMethod() == 0 {
-			result.Set(reflect.ValueOf(value))
-			return newOffset, nil
-		}
-	}
-	return newOffset, mmdberrors.NewUnmarshalTypeError(value, result.Type())
-}
-
-func (d *Decoder) unmarshalInt32(size, offset uint, result reflect.Value) (uint, error) {
-	if size > 4 {
-		return 0, mmdberrors.NewInvalidDatabaseError(
-			"the MaxMind DB file's data section contains bad data (int32 size of %v)",
-			size,
-		)
-	}
-	value, newOffset := d.decodeInt(size, offset)
-
-	switch result.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		n := int64(value)
-		if !result.OverflowInt(n) {
-			result.SetInt(n)
-			return newOffset, nil
-		}
-	case reflect.Uint,
-		reflect.Uint8,
-		reflect.Uint16,
-		reflect.Uint32,
-		reflect.Uint64,
-		reflect.Uintptr:
-		n := uint64(value)
-		if !result.OverflowUint(n) {
-			result.SetUint(n)
-			return newOffset, nil
-		}
-	case reflect.Interface:
-		if result.NumMethod() == 0 {
-			result.Set(reflect.ValueOf(value))
-			return newOffset, nil
-		}
-	}
-	return newOffset, mmdberrors.NewUnmarshalTypeError(value, result.Type())
-}
-
-func (d *Decoder) unmarshalMap(
-	size uint,
-	offset uint,
-	result reflect.Value,
-	depth int,
-) (uint, error) {
-	result = indirect(result)
-	switch result.Kind() {
-	default:
-		return 0, mmdberrors.NewUnmarshalTypeStrError("map", result.Type())
-	case reflect.Struct:
-		return d.decodeStruct(size, offset, result, depth)
-	case reflect.Map:
-		return d.decodeMap(size, offset, result, depth)
-	case reflect.Interface:
-		if result.NumMethod() == 0 {
-			rv := reflect.ValueOf(make(map[string]any, size))
-			newOffset, err := d.decodeMap(size, offset, rv, depth)
-			result.Set(rv)
-			return newOffset, err
-		}
-		return 0, mmdberrors.NewUnmarshalTypeStrError("map", result.Type())
-	}
-}
-
-func (d *Decoder) unmarshalPointer(
-	size, offset uint,
-	result reflect.Value,
-	depth int,
-) (uint, error) {
-	pointer, newOffset, err := d.decodePointer(size, offset)
-	if err != nil {
-		return 0, err
-	}
-	_, err = d.decode(pointer, result, depth)
-	return newOffset, err
-}
-
-func (d *Decoder) unmarshalSlice(
-	size uint,
-	offset uint,
-	result reflect.Value,
-	depth int,
-) (uint, error) {
-	switch result.Kind() {
-	case reflect.Slice:
-		return d.decodeSlice(size, offset, result, depth)
-	case reflect.Interface:
-		if result.NumMethod() == 0 {
-			a := []any{}
-			rv := reflect.ValueOf(&a).Elem()
-			newOffset, err := d.decodeSlice(size, offset, rv, depth)
-			result.Set(rv)
-			return newOffset, err
-		}
-	}
-	return 0, mmdberrors.NewUnmarshalTypeStrError("array", result.Type())
-}
-
-func (d *Decoder) unmarshalString(size, offset uint, result reflect.Value) (uint, error) {
-	value, newOffset := d.decodeString(size, offset)
-
-	switch result.Kind() {
-	case reflect.String:
-		result.SetString(value)
-		return newOffset, nil
-	case reflect.Interface:
-		if result.NumMethod() == 0 {
-			result.Set(reflect.ValueOf(value))
-			return newOffset, nil
-		}
-	}
-	return newOffset, mmdberrors.NewUnmarshalTypeError(value, result.Type())
-}
-
-func (d *Decoder) unmarshalUint(
-	size, offset uint,
-	result reflect.Value,
-	uintType uint,
-) (uint, error) {
-	if size > uintType/8 {
-		return 0, mmdberrors.NewInvalidDatabaseError(
-			"the MaxMind DB file's data section contains bad data (uint%v size of %v)",
-			uintType,
-			size,
-		)
-	}
-
-	value, newOffset := d.decodeUint(size, offset)
-
-	switch result.Kind() {
-	case reflect.Int, reflect.Int8, reflect.Int16, reflect.Int32, reflect.Int64:
-		n := int64(value)
-		if !result.OverflowInt(n) {
-			result.SetInt(n)
-			return newOffset, nil
-		}
-	case reflect.Uint,
-		reflect.Uint8,
-		reflect.Uint16,
-		reflect.Uint32,
-		reflect.Uint64,
-		reflect.Uintptr:
-		if !result.OverflowUint(value) {
-			result.SetUint(value)
-			return newOffset, nil
-		}
-	case reflect.Interface:
-		if result.NumMethod() == 0 {
-			result.Set(reflect.ValueOf(value))
-			return newOffset, nil
-		}
-	}
-	return newOffset, mmdberrors.NewUnmarshalTypeError(value, result.Type())
-}
-
-var bigIntType = reflect.TypeOf(big.Int{})
-
-func (d *Decoder) unmarshalUint128(size, offset uint, result reflect.Value) (uint, error) {
-	if size > 16 {
-		return 0, mmdberrors.NewInvalidDatabaseError(
-			"the MaxMind DB file's data section contains bad data (uint128 size of %v)",
-			size,
-		)
-	}
-	value, newOffset := d.decodeUint128(size, offset)
-
-	switch result.Kind() {
-	case reflect.Struct:
-		if result.Type() == bigIntType {
-			result.Set(reflect.ValueOf(*value))
-			return newOffset, nil
-		}
-	case reflect.Interface:
-		if result.NumMethod() == 0 {
-			result.Set(reflect.ValueOf(value))
-			return newOffset, nil
-		}
-	}
-	return newOffset, mmdberrors.NewUnmarshalTypeError(value, result.Type())
-}
-
 func decodeBool(size, offset uint) (bool, uint) {
 	return size != 0, offset
 }
 
-func (d *Decoder) decodeBytes(size, offset uint) ([]byte, uint) {
+func (d *DataDecoder) decodeBytes(size, offset uint) ([]byte, uint) {
 	newOffset := offset + size
 	bytes := make([]byte, size)
 	copy(bytes, d.buffer[offset:newOffset])
 	return bytes, newOffset
 }
 
-func (d *Decoder) decodeFloat64(size, offset uint) (float64, uint) {
+func (d *DataDecoder) decodeFloat64(size, offset uint) (float64, uint) {
 	newOffset := offset + size
 	bits := binary.BigEndian.Uint64(d.buffer[offset:newOffset])
 	return math.Float64frombits(bits), newOffset
 }
 
-func (d *Decoder) decodeFloat32(size, offset uint) (float32, uint) {
+func (d *DataDecoder) decodeFloat32(size, offset uint) (float32, uint) {
 	newOffset := offset + size
 	bits := binary.BigEndian.Uint32(d.buffer[offset:newOffset])
 	return math.Float32frombits(bits), newOffset
 }
 
-func (d *Decoder) decodeInt(size, offset uint) (int, uint) {
+func (d *DataDecoder) decodeInt(size, offset uint) (int, uint) {
 	newOffset := offset + size
 	var val int32
 	for _, b := range d.buffer[offset:newOffset] {
@@ -702,46 +227,7 @@ func (d *Decoder) decodeInt(size, offset uint) (int, uint) {
 	return int(val), newOffset
 }
 
-func (d *Decoder) decodeMap(
-	size uint,
-	offset uint,
-	result reflect.Value,
-	depth int,
-) (uint, error) {
-	if result.IsNil() {
-		result.Set(reflect.MakeMapWithSize(result.Type(), int(size)))
-	}
-
-	mapType := result.Type()
-	keyValue := reflect.New(mapType.Key()).Elem()
-	elemType := mapType.Elem()
-	var elemValue reflect.Value
-	for range size {
-		var key []byte
-		var err error
-		key, offset, err = d.decodeKey(offset)
-		if err != nil {
-			return 0, err
-		}
-
-		if elemValue.IsValid() {
-			elemValue.SetZero()
-		} else {
-			elemValue = reflect.New(elemType).Elem()
-		}
-
-		offset, err = d.decode(offset, elemValue, depth)
-		if err != nil {
-			return 0, fmt.Errorf("decoding value for %s: %w", key, err)
-		}
-
-		keyValue.SetString(string(key))
-		result.SetMapIndex(keyValue, elemValue)
-	}
-	return offset, nil
-}
-
-func (d *Decoder) decodeMapToDeserializer(
+func (d *DataDecoder) decodeMapToDeserializer(
 	size uint,
 	offset uint,
 	dser deserializer,
@@ -770,7 +256,7 @@ func (d *Decoder) decodeMapToDeserializer(
 	return offset, nil
 }
 
-func (d *Decoder) decodePointer(
+func (d *DataDecoder) decodePointer(
 	size uint,
 	offset uint,
 ) (uint, uint, error) {
@@ -805,24 +291,7 @@ func (d *Decoder) decodePointer(
 	return pointer, newOffset, nil
 }
 
-func (d *Decoder) decodeSlice(
-	size uint,
-	offset uint,
-	result reflect.Value,
-	depth int,
-) (uint, error) {
-	result.Set(reflect.MakeSlice(result.Type(), int(size), int(size)))
-	for i := range size {
-		var err error
-		offset, err = d.decode(offset, result.Index(int(i)), depth)
-		if err != nil {
-			return 0, err
-		}
-	}
-	return offset, nil
-}
-
-func (d *Decoder) decodeSliceToDeserializer(
+func (d *DataDecoder) decodeSliceToDeserializer(
 	size uint,
 	offset uint,
 	dser deserializer,
@@ -845,95 +314,12 @@ func (d *Decoder) decodeSliceToDeserializer(
 	return offset, nil
 }
 
-func (d *Decoder) decodeString(size, offset uint) (string, uint) {
+func (d *DataDecoder) decodeString(size, offset uint) (string, uint) {
 	newOffset := offset + size
 	return string(d.buffer[offset:newOffset]), newOffset
 }
 
-func (d *Decoder) decodeStruct(
-	size uint,
-	offset uint,
-	result reflect.Value,
-	depth int,
-) (uint, error) {
-	fields := cachedFields(result)
-
-	// This fills in embedded structs
-	for _, i := range fields.anonymousFields {
-		_, err := d.unmarshalMap(size, offset, result.Field(i), depth)
-		if err != nil {
-			return 0, err
-		}
-	}
-
-	// This handles named fields
-	for range size {
-		var (
-			err error
-			key []byte
-		)
-		key, offset, err = d.decodeKey(offset)
-		if err != nil {
-			return 0, err
-		}
-		// The string() does not create a copy due to this compiler
-		// optimization: https://github.com/golang/go/issues/3512
-		j, ok := fields.namedFields[string(key)]
-		if !ok {
-			offset, err = d.nextValueOffset(offset, 1)
-			if err != nil {
-				return 0, err
-			}
-			continue
-		}
-
-		offset, err = d.decode(offset, result.Field(j), depth)
-		if err != nil {
-			return 0, fmt.Errorf("decoding value for %s: %w", key, err)
-		}
-	}
-	return offset, nil
-}
-
-type fieldsType struct {
-	namedFields     map[string]int
-	anonymousFields []int
-}
-
-var fieldsMap sync.Map
-
-func cachedFields(result reflect.Value) *fieldsType {
-	resultType := result.Type()
-
-	if fields, ok := fieldsMap.Load(resultType); ok {
-		return fields.(*fieldsType)
-	}
-	numFields := resultType.NumField()
-	namedFields := make(map[string]int, numFields)
-	var anonymous []int
-	for i := range numFields {
-		field := resultType.Field(i)
-
-		fieldName := field.Name
-		if tag := field.Tag.Get("maxminddb"); tag != "" {
-			if tag == "-" {
-				continue
-			}
-			fieldName = tag
-		}
-		if field.Anonymous {
-			anonymous = append(anonymous, i)
-			continue
-		}
-		namedFields[fieldName] = i
-	}
-	fields := &fieldsType{namedFields, anonymous}
-	fieldsMap.Store(resultType, fields)
-
-	return fields
-}
-
-func (d *Decoder) decodeUint(size, offset uint) (uint64, uint) {
+func (d *DataDecoder) decodeUint(size, offset uint) (uint64, uint) {
 	newOffset := offset + size
 	bytes := d.buffer[offset:newOffset]
 
@@ -944,7 +330,7 @@ func (d *Decoder) decodeUint(size, offset uint) (uint64, uint) {
 	return val, newOffset
 }
 
-func (d *Decoder) decodeUint128(size, offset uint) (*big.Int, uint) {
+func (d *DataDecoder) decodeUint128(size, offset uint) (*big.Int, uint) {
 	newOffset := offset + size
 	val := new(big.Int)
 	val.SetBytes(d.buffer[offset:newOffset])
@@ -964,7 +350,7 @@ func uintFromBytes(prefix uint, uintBytes []byte) uint {
 // can take advantage of https://github.com/golang/go/issues/3512 to avoid
 // copying the bytes when decoding a struct. Previously, we achieved this by
 // using unsafe.
-func (d *Decoder) decodeKey(offset uint) ([]byte, uint, error) {
+func (d *DataDecoder) decodeKey(offset uint) ([]byte, uint, error) {
 	typeNum, size, dataOffset, err := d.decodeCtrlData(offset)
 	if err != nil {
 		return nil, 0, err
@@ -993,7 +379,7 @@ func (d *Decoder) decodeKey(offset uint) ([]byte, uint, error) {
 // This function is used to skip ahead to the next value without decoding
 // the one at the offset passed in. The size bits have different meanings for
 // different data types.
-func (d *Decoder) nextValueOffset(offset, numberToSkip uint) (uint, error) {
+func (d *DataDecoder) nextValueOffset(offset, numberToSkip uint) (uint, error) {
 	if numberToSkip == 0 {
 		return offset, nil
 	}
