+++ internal/decoder/decoder_test.go
@@ -208,7 +208,7 @@ func validateDecoding(t *testing.T, tests map[string]any) {
 	for inputStr, expected := range tests {
 		inputBytes, err := hex.DecodeString(inputStr)
 		require.NoError(t, err)
-		d := Decoder{buffer: inputBytes}
+		d := New(inputBytes)
 
 		var result any
 		_, err = d.decode(0, reflect.ValueOf(&result), 0)
@@ -224,7 +224,7 @@ func validateDecoding(t *testing.T, tests map[string]any) {
 func TestPointers(t *testing.T) {
 	bytes, err := os.ReadFile(testFile("maps-with-pointers.raw"))
 	require.NoError(t, err)
-	d := Decoder{buffer: bytes}
+	d := New(bytes)
 
 	expected := map[uint]map[string]string{
 		0:  {"long_key": "long_value1"},
